#60:  app module의 SplashActivity 수정 및 SharedPreferences 설정 추가
