+++ app/src/main/java/com/jyp/journeypiki/presentation/SplashScreen.kt
@@ -1,4 +1,4 @@
-package com.jyp.journeypiki.presentation.splash
+package com.jyp.journeypiki.presentation
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,14 +8,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 
+
 @Composable
 fun SplashScreen() {
     Box(
-            modifier = Modifier.fillMaxSize(),
+        modifier = Modifier.fillMaxSize(),
     ) {
         Text(
-                modifier = Modifier.align(Alignment.Center),
-                text = "splash",
+            modifier = Modifier.align(Alignment.Center),
+            text = "splash",
         )
     }
 }
